Abstracted the TestWatcher methods
I abstracted the TestWatcher methods so that they can be overridden
from individual tests.
Fixed empty deviceName bug.
