+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.joss</groupId>
     <artifactId>conductor-mobile</artifactId>
-    <version>0.5.1</version>
+    <version>0.6.0</version>
 
     <repositories>
         <repository>
@@ -14,6 +14,10 @@
             <url>https://jitpack.io</url>
         </repository>
     </repositories>
+
+    <scm>
+        <url>https://github.com/jossjacobo/conductor-mobile</url>
+    </scm>
 
     <dependencies>
         <dependency>
@@ -59,4 +63,44 @@
         </dependency>
 
     </dependencies>
+
+    <licenses>
+        <license>
+            <name>Apache License, Version 2.0</name>
+            <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
+            <distribution>source</distribution>
+        </license>
+    </licenses>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-source-plugin</artifactId>
+                <version>2.2.1</version>
+                <executions>
+                    <execution>
+                        <id>attach-sources</id>
+                        <goals>
+                            <goal>jar-no-fork</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-javadoc-plugin</artifactId>
+                <version>2.9.1</version>
+                <executions>
+                    <execution>
+                        <id>attach-javadocs</id>
+                        <goals>
+                            <goal>jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
+
 </project>
