+++ src/main/java/com/joss/conductor/mobile/Watchman.java
@@ -1,0 +1,53 @@
+package com.joss.conductor.mobile;
+
+import com.joss.conductor.mobile.util.ScreenShotUtil;
+import org.junit.After;
+import org.junit.Rule;
+import org.junit.rules.TestRule;
+import org.junit.rules.TestWatcher;
+import org.junit.runner.Description;
+
+/**
+ * Created on 2/10/17.
+ */
+public abstract class Watchman extends TestWatcher {
+
+    @Rule
+    public TestRule watchman = this;
+    public abstract Locomotive getLocomotive();
+
+    private boolean failure;
+    private Throwable e;
+
+    @Override
+    protected void failed(Throwable e, Description description) {
+        if (getLocomotive().configuration.screenshotsOnFail()) {
+            failure = true;
+            this.e = e;
+        }
+    }
+
+    /**
+     * Take screenshot if the test failed.
+     */
+    @Override
+    protected void finished(Description description) {
+        super.finished(description);
+        if (getLocomotive().configuration.screenshotsOnFail()) {
+            if (failure) {
+                ScreenShotUtil.take(getLocomotive(),
+                        description.getDisplayName(),
+                        e.getMessage());
+            }
+            getLocomotive().driver.quit();
+        }
+    }
+
+    @After
+    public void teardown() {
+        if (!getLocomotive().configuration.screenshotsOnFail()) {
+            getLocomotive().driver.quit();
+        }
+    }
+
+}
