+++ src/main/java/com/joss/conductor/mobile/util/PropertiesUtil.java
@@ -1,6 +1,7 @@
 package com.joss.conductor.mobile.util;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.util.Properties;
 
 /**
@@ -10,10 +11,13 @@ public class PropertiesUtil {
 
     private static final String DEFAULT_PROPERTIES_PATH = "/default.properties";
 
-    public static Properties getDefaultProperties(Object o) {
+    public static Properties getDefaultProperties(Object object) {
         Properties props = new Properties();
         try {
-            props.load(o.getClass().getResourceAsStream(DEFAULT_PROPERTIES_PATH));
+            InputStream inputStream = object.getClass().getResourceAsStream(DEFAULT_PROPERTIES_PATH);
+            if (inputStream != null) {
+                props.load(inputStream);
+            }
         } catch (IOException ioException) {
             ioException.printStackTrace();
             Log.fatal("Couldn\'t load in default properties");
